requirements.txt
pymongo
redis
pytest
requests
selenium
aiohttp
lxml
beautifulsoup4
pyquery
tesserocr
pillow
pyspider
Scrapy
scrapy-splash
scrapy-redis
scrapyd-client
python-scrapyd-api
scrapyrt